automatic Sys11 and DE matrix numbering support
This is the first version with "automatic" System 11 and Data East support in your VBScript for the table, also set the Controller's new Sys11 property as:
         controller.Sys11=True
this will have the VPCOM bridge auto-renumber lamps and switches from the 1-64 numbering in the manual (and that all Sys11/DE VPT tables use) to the P-ROC expected Col/Row numbering.  No more renumbering all the lamps and switches in the VBScript for System11 and Data East machines.  Tested on DE Hook; more bugs may be hiding...